Rename TypicalPersons to TypicalEntities
TypicalPersons came from AB3. Now we're using
both Projects and Clients
