+++ src/test/java/mycelium/mycelium/logic/commands/ClearCommandTest.java
@@ -1,7 +1,7 @@
 package mycelium.mycelium.logic.commands;
 
 import static mycelium.mycelium.logic.commands.CommandTestUtil.assertCommandSuccess;
-import static mycelium.mycelium.testutil.TypicalPersons.getTypicalAddressBook;
+import static mycelium.mycelium.testutil.TypicalEntities.getTypicalAddressBook;
 
 import org.junit.jupiter.api.Test;
