+++ src/test/java/mycelium/mycelium/logic/commands/DeleteCommandTest.java
@@ -5,7 +5,7 @@ import static mycelium.mycelium.logic.commands.CommandTestUtil.assertCommandSucc
 import static mycelium.mycelium.logic.commands.CommandTestUtil.showPersonAtIndex;
 import static mycelium.mycelium.testutil.TypicalIndexes.INDEX_FIRST_PERSON;
 import static mycelium.mycelium.testutil.TypicalIndexes.INDEX_SECOND_PERSON;
-import static mycelium.mycelium.testutil.TypicalPersons.getTypicalAddressBook;
+import static mycelium.mycelium.testutil.TypicalEntities.getTypicalAddressBook;
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
