+++ src/test/java/mycelium/mycelium/logic/commands/ListCommandTest.java
@@ -3,7 +3,7 @@ package mycelium.mycelium.logic.commands;
 import static mycelium.mycelium.logic.commands.CommandTestUtil.assertCommandSuccess;
 import static mycelium.mycelium.logic.commands.CommandTestUtil.showPersonAtIndex;
 import static mycelium.mycelium.testutil.TypicalIndexes.INDEX_FIRST_PERSON;
-import static mycelium.mycelium.testutil.TypicalPersons.getTypicalAddressBook;
+import static mycelium.mycelium.testutil.TypicalEntities.getTypicalAddressBook;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
