+++ src/test/java/mycelium/mycelium/model/AddressBookTest.java
@@ -3,9 +3,9 @@ package mycelium.mycelium.model;
 import static mycelium.mycelium.logic.commands.CommandTestUtil.VALID_ADDRESS_BOB;
 import static mycelium.mycelium.logic.commands.CommandTestUtil.VALID_TAG_HUSBAND;
 import static mycelium.mycelium.testutil.Assert.assertThrows;
-import static mycelium.mycelium.testutil.TypicalPersons.ALICE;
-import static mycelium.mycelium.testutil.TypicalPersons.WEST;
-import static mycelium.mycelium.testutil.TypicalPersons.getTypicalAddressBook;
+import static mycelium.mycelium.testutil.TypicalEntities.ALICE;
+import static mycelium.mycelium.testutil.TypicalEntities.WEST;
+import static mycelium.mycelium.testutil.TypicalEntities.getTypicalAddressBook;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertNotEquals;
