+++ src/test/java/mycelium/mycelium/model/person/UniquePersonListTest.java
@@ -3,8 +3,8 @@ package mycelium.mycelium.model.person;
 import static mycelium.mycelium.logic.commands.CommandTestUtil.VALID_ADDRESS_BOB;
 import static mycelium.mycelium.logic.commands.CommandTestUtil.VALID_TAG_HUSBAND;
 import static mycelium.mycelium.testutil.Assert.assertThrows;
-import static mycelium.mycelium.testutil.TypicalPersons.ALICE;
-import static mycelium.mycelium.testutil.TypicalPersons.BOB;
+import static mycelium.mycelium.testutil.TypicalEntities.ALICE;
+import static mycelium.mycelium.testutil.TypicalEntities.BOB;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
