+++ src/test/java/mycelium/mycelium/storage/JsonAdaptedClientTest.java
@@ -1,7 +1,7 @@
 package mycelium.mycelium.storage;
 
 import static mycelium.mycelium.testutil.Assert.assertThrows;
-import static mycelium.mycelium.testutil.TypicalPersons.WEST;
+import static mycelium.mycelium.testutil.TypicalEntities.WEST;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
 import java.io.IOException;
